+++ gcm_agent/utils/__init__.py
@@ -2,6 +2,7 @@
 
 # Made with Bob
 # 2026-06-05 19:54 UTC - Added exports for logger utilities
+# 2026-06-09 21:24 UTC - Added table formatter export
 
 from gcm_agent.utils.logger import (
     get_logger,
@@ -13,6 +14,7 @@ from gcm_agent.utils.logger import (
     get_agent_logger,
     get_ui_logger,
 )
+from gcm_agent.utils.table_formatter import format_response_tables
 
 __all__ = [
     "get_logger",
@@ -23,4 +25,5 @@ __all__ = [
     "get_mcp_logger",
     "get_agent_logger",
     "get_ui_logger",
+    "format_response_tables",
 ]